+++ plugins/gs/src/components/MainMenu/MainMenu.tsx
@@ -3,6 +3,7 @@ import { SidebarGroup, SidebarItem } from '@backstage/core-components';
 import { FeatureFlagged } from '@backstage/core-app-api';
 import { FeatureEnabled } from '../FeatureEnabled';
 import { GiantSwarmIcon } from '../../assets/icons/CustomIcons';
+import PlaceIcon from '@material-ui/icons/Place';
 
 export const MainMenu = () => {
   return (
@@ -10,7 +11,7 @@ export const MainMenu = () => {
       <FeatureEnabled feature="installationsPage">
         <FeatureFlagged with="show-installations-page">
           <SidebarItem
-            icon={GiantSwarmIcon}
+            icon={PlaceIcon}
             to="installations"
             text="Installations"
           />