+++ plugins/gs/src/components/catalog/columns.tsx
@@ -229,6 +229,11 @@ export const columnFactories = Object.freeze({
       filtering: false,
       render: ({ entity }) =>
         entity.metadata?.labels?.['giantswarm.io/customer'] ?? '',
+      customFilterAndSearch(query: string, { entity }) {
+        return formatContent(entity)
+          .toLowerCase()
+          .includes(query.toLowerCase());
+      },
       customSort({ entity: entity1 }, { entity: entity2 }) {
         return formatContent(entity1).localeCompare(formatContent(entity2));
       },
@@ -244,6 +249,11 @@ export const columnFactories = Object.freeze({
       width: 'auto',
       filtering: false,
       render: ({ entity }) => formatContent(entity),
+      customFilterAndSearch(query: string, { entity }) {
+        return formatContent(entity)
+          .toLowerCase()
+          .includes(query.toLowerCase());
+      },
       customSort({ entity: entity1 }, { entity: entity2 }) {
         return formatContent(entity1).localeCompare(formatContent(entity2));
       },
@@ -259,6 +269,11 @@ export const columnFactories = Object.freeze({
       width: 'auto',
       filtering: false,
       render: ({ entity }) => formatContent(entity),
+      customFilterAndSearch(query: string, { entity }) {
+        return formatContent(entity)
+          .toLowerCase()
+          .includes(query.toLowerCase());
+      },
       customSort({ entity: entity1 }, { entity: entity2 }) {
         return formatContent(entity1).localeCompare(formatContent(entity2));
       },
@@ -274,6 +289,53 @@ export const columnFactories = Object.freeze({
       width: 'auto',
       filtering: false,
       render: ({ entity }) => formatContent(entity),
+      customFilterAndSearch(query: string, { entity }) {
+        return formatContent(entity)
+          .toLowerCase()
+          .includes(query.toLowerCase());
+      },
+      customSort({ entity: entity1 }, { entity: entity2 }) {
+        return formatContent(entity1).localeCompare(formatContent(entity2));
+      },
+    };
+  },
+  createBaseColumn(): TableColumn<CatalogTableRow> {
+    function formatContent(entity: Entity): string {
+      return entity.metadata?.annotations?.['giantswarm.io/base'] ?? '';
+    }
+
+    return {
+      title: 'Base domain',
+      width: 'auto',
+      filtering: false,
+      render: ({ entity }) => formatContent(entity),
+      customFilterAndSearch(query: string, { entity }) {
+        return formatContent(entity)
+          .toLowerCase()
+          .includes(query.toLowerCase());
+      },
+      customSort({ entity: entity1 }, { entity: entity2 }) {
+        return formatContent(entity1).localeCompare(formatContent(entity2));
+      },
+    };
+  },
+  createAccountEngineerColumn(): TableColumn<CatalogTableRow> {
+    function formatContent(entity: Entity): string {
+      return (
+        entity.metadata?.annotations?.['giantswarm.io/account-engineer'] ?? ''
+      );
+    }
+
+    return {
+      title: 'Account engineer',
+      width: 'auto',
+      filtering: false,
+      render: ({ entity }) => formatContent(entity),
+      customFilterAndSearch(query: string, { entity }) {
+        return formatContent(entity)
+          .toLowerCase()
+          .includes(query.toLowerCase());
+      },
       customSort({ entity: entity1 }, { entity: entity2 }) {
         return formatContent(entity1).localeCompare(formatContent(entity2));
       },